src/requirements.txt
fastapi
uvicorn
databases[sqlite]
pyjwt
python-multipart
requests